requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pillow
numpy
matplotlib
imageio
imageio-ffmpeg
